pyproject.toml
[project]
name = "metaso-cli"
version = "0.1.0"
description = "CLI tool for calling LLM APIs with streaming support"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]
